+++ samples/constitutions/default/actions.js
@@ -269,7 +269,12 @@ function checkReconfigurationType(config, new_config) {
   const from = config.reconfiguration_type;
   const to = new_config.reconfiguration_type;
   if (from !== to && to !== undefined) {
-    if (!(from === "OneTransaction" && to === "TwoTransaction")) {
+    if (
+      !(
+        (from === undefined || from === "OneTransaction") &&
+        to === "TwoTransaction"
+      )
+    ) {
       throw new Error(
         `Cannot change reconfiguration type from ${from} to ${to}.`
       );
@@ -940,41 +945,6 @@ const actions = new Map([
               ccf.strToBuf(endorsed_node_cert)
             );
           }
-
-          if (serviceConfig.reconfiguration_type == "TwoTransaction") {
-            const latest_id_raw = ccf.kv[
-              "public:ccf.gov.nodes.network.configurations"
-            ].get(getSingletonKvKey());
-            if (latest_id_raw === undefined) {
-              throw new Error("Network configuration could not be found");
-            }
-            const latest_id = ccf.bufToJsonCompatible(latest_id_raw);
-            const rid_buf = new ArrayBuffer(8);
-            new DataView(rid_buf).setUint32(0, latest_id.rid, true);
-
-            const latest_config_raw =
-              ccf.kv["public:ccf.gov.nodes.network.configurations"].get(
-                rid_buf
-              );
-            if (latest_config_raw === undefined) {
-              throw new Error("Network configuration could not be found");
-            }
-            const latest_config = ccf.bufToJsonCompatible(latest_config_raw);
-
-            latest_config.nodes.push(args.node_id);
-            latest_config.rid++;
-            new DataView(rid_buf).setUint32(0, latest_config.rid, true);
-            ccf.kv["public:ccf.gov.nodes.network.configurations"].set(
-              rid_buf,
-              ccf.jsonCompatibleToBuf(latest_config)
-            );
-            latest_config.nodes = {};
-            latest_id.rid = latest_config.rid;
-            ccf.kv["public:ccf.gov.nodes.network.configurations"].set(
-              getSingletonKvKey(),
-              ccf.jsonCompatibleToBuf(latest_id)
-            );
-          }
         }
       }
     ),
@@ -1008,53 +978,23 @@ const actions = new Map([
         const node = ccf.kv["public:ccf.gov.nodes.info"].get(
           ccf.strToBuf(args.node_id)
         );
-        if (node !== undefined) {
-          const node_obj = ccf.bufToJsonCompatible(node);
+        if (node === undefined) {
+          return;
+        }
+        const node_obj = ccf.bufToJsonCompatible(node);
+        if (node_obj.status === "Pending") {
+          ccf.kv["public:ccf.gov.nodes.info"].delete(
+            ccf.strToBuf(args.node_id)
+          );
+        } else {
           node_obj.status =
-            serviceConfig.reconfiguration_type === "TwoTransaction" &&
-            node_obj.status !== "Pending"
+            serviceConfig.reconfiguration_type === "TwoTransaction"
               ? "Retiring"
               : "Retired";
           ccf.kv["public:ccf.gov.nodes.info"].set(
             ccf.strToBuf(args.node_id),
             ccf.jsonCompatibleToBuf(node_obj)
           );
-
-          if (serviceConfig.reconfiguration_type == "TwoTransaction") {
-            const latest_id_raw = ccf.kv[
-              "public:ccf.gov.nodes.network.configurations"
-            ].get(getSingletonKvKey());
-            if (latest_id_raw === undefined) {
-              throw new Error("Network configuration could not be found");
-            }
-            const latest_id = ccf.bufToJsonCompatible(latest_id_raw);
-            const rid_buf = new ArrayBuffer(8);
-            new DataView(rid_buf).setUint32(0, latest_id.rid, true);
-            const latest_config_raw =
-              ccf.kv["public:ccf.gov.nodes.network.configurations"].get(
-                rid_buf
-              );
-            if (latest_config_raw === undefined) {
-              throw new Error("Network configuration could not be found");
-            }
-            const latest_config = ccf.bufToJsonCompatible(latest_config_raw);
-            const idx = latest_config.nodes.indexOf(args.node_id);
-            if (idx > -1) {
-              latest_config.nodes.splice(idx, 1);
-            }
-            latest_config.rid++;
-            new DataView(rid_buf).setUint32(0, latest_config.rid, true);
-            ccf.kv["public:ccf.gov.nodes.network.configurations"].set(
-              rid_buf,
-              ccf.jsonCompatibleToBuf(latest_config)
-            );
-            latest_config.nodes = {};
-            latest_id.rid = latest_config.rid;
-            ccf.kv["public:ccf.gov.nodes.network.configurations"].set(
-              getSingletonKvKey(),
-              ccf.jsonCompatibleToBuf(latest_id)
-            );
-          }
         }
       }
     ),